+++ test/middleware-config/pages/+Head.tsx
@@ -1,0 +1,12 @@
+// https://vike.dev/Head
+
+import React from 'react'
+import logoUrl from '../assets/logo.svg'
+
+export default function HeadDefault() {
+  return (
+    <>
+      <link rel="icon" href={logoUrl} />
+    </>
+  )
+}