+++ test/middleware-config/pages/Middleware.ts
@@ -1,0 +1,9 @@
+export { Middleware }
+
+import type { UniversalMiddleware } from '@universal-middleware/core'
+
+type Middleware = {
+  name: string
+  order: 'pre' | 'post' | number
+  value: UniversalMiddleware
+}