+++ test/middleware-config/pages/index/Counter.tsx
@@ -1,0 +1,11 @@
+import React, { useState } from 'react'
+
+export function Counter() {
+  const [count, setCount] = useState(0)
+
+  return (
+    <button type="button" onClick={() => setCount((count) => count + 1)}>
+      Counter {count}
+    </button>
+  )
+}